Materials applied
All materials have been re-imported and re-named to "kidLinda" so we can differentiate between Linda and Baby's (dolly) materials. The UV's for the pigtails and hairBack may be the previous version, they don't match the new materials made and I am not sure if I can move the UV's (I don't remember if that would effect the rig). Can adjust or work that out on Monday.